+++ data/duke.txt
@@ -1,3 +1,3 @@
-  | X | return book  | 2019-10-15T18:00
+  |   | return book  | 2019-10-15T18:00
   |   | return book  | 2019-10-15T18:00
   |   | hello
