+++ src/main/java/duke/exceptions/StorageException.java
@@ -1,0 +1,8 @@
+package duke.exceptions;
+
+public class StorageException extends DukeException {
+    public StorageException(String message) {
+        super(message);
+    }
+
+}
