+++ src/main/java/duke/main/Parser.java
@@ -1,0 +1,4 @@
+package duke.main;
+
+public class Parser {
+}
