+++ src/main/java/duke/main/TaskList.java
@@ -1,0 +1,13 @@
+package duke.main;
+
+import duke.task.Task;
+import java.util.ArrayList;
+
+public class TaskList {
+    private ArrayList<Task> tasks;
+
+    public TaskList() {
+        tasks = new ArrayList<>();
+    }
+
+}
