+++ src/main/java/duke/main/Ui.java
@@ -1,0 +1,21 @@
+package duke.main;
+
+public class Ui {
+    public Ui() {
+    }
+
+    public void welcome() {
+        String logo = " ____        _        \n"
+                + "|  _ \\ _   _| | _____ \n"
+                + "| | | | | | | |/ / _ \\\n"
+                + "| |_| | |_| |   <  __/\n"
+                + "|____/ \\__,_|_|\\_\\___|\n";
+        System.out.println("Hello from\n" + logo);
+        System.out.println("Hello! I'm SUPERRRRDUKE");
+        System.out.println("What can I do for you?");
+    }
+
+    public void bye() {
+        System.out.println("Bye. Hope to see you again soon!");
+    }
+}
