+++ src/main/java/duke/task/TaskType.java
@@ -1,0 +1,5 @@
+package duke.task;
+
+public enum TaskType {
+    TODO, EVENT, DEADLINE;
+}
