+++ src/main/java/duke/task/Todo.java
@@ -1,0 +1,17 @@
+package duke.task;
+
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description, TODO);
+    }
+
+    @Override
+    public String toString() {
+        return super.getSymbol() + super.toString();
+    }
+
+    @Override
+    public String convertToSaveFormat() {
+        return String.format("%s", super.convertToSaveFormat());
+    }
+}
